moved function outside of app and into question div component
