+++ src/QuestionBox.js
@@ -1,20 +1,27 @@
 import Answers from './Answers'
-import App from './App'
 
 const QuestionBox = ({ questions, questionNumber, answers, handleAnswerClick }) => {
+  const stringFormatter = (str) => {
+    let string = str.replace(/&quot;/g, '\"')
+    string = string.replace(/&#039;/g, '\'')
+    string = string.replace(/&amp;/g, '&')
+  
+    return string
+  }
+
 	return (
 		<>
 			<h3 className='question'>Question {questionNumber + 1} out of {questions.data.length}</h3>
-			<h5 className='card-title card-text'>{App.stringFormatter(questions.data[questionNumber].question)}</h5>
+			<h5 className='card-title card-text'>{stringFormatter(questions.data[questionNumber].question)}</h5>
 			<ul>
 				{
 					answers.map((option, index) => {
 						return (
 							<Answers
 								key={index}
-								answer={App.stringFormatter(option)}
+								answer={stringFormatter(option)}
 								onSubmit={handleAnswerClick}
-								correct={App.stringFormatter(questions.data[questionNumber].correct_answer)}
+								correct={stringFormatter(questions.data[questionNumber].correct_answer)}
 							></Answers>
 						)
 					})
